issue #121 move tests to proper packages + remove non-test
